refactor: integrate CrmDataProvider into CRM pages and remove unused route utility
