+++ frontend/pages/_app.tsx
@@ -20,7 +20,6 @@ import { theme } from "~community/common/theme/theme";
 import { themeSelector } from "~community/common/theme/themeSelector";
 import { MyAppPropsType } from "~community/common/types/CommonTypes";
 import { getDataFromLocalStorage } from "~community/common/utils/accessLocalStorage";
-import { CrmDataProvider } from "~community/crm/v2/providers/CrmDataProvider";
 import "~enterprise/common/components/atoms/driverJsPopover/styles.css";
 import AnnouncementWrapper from "~enterprise/common/components/organisms/AnnouncementWrapper/AnnouncementWrapper";
 import {
@@ -104,11 +103,9 @@ function MyApp({
                 <ThemeProvider theme={newTheme}>
                   <I18nextProvider i18n={i18n}>
                     <ErrorBoundary FallbackComponent={Error}>
-                      <CrmDataProvider>
-                        <BaseLayout>
-                          <Component {...pageProps} />
-                        </BaseLayout>
-                      </CrmDataProvider>
+                      <BaseLayout>
+                        <Component {...pageProps} />
+                      </BaseLayout>
                     </ErrorBoundary>
                     <ReactQueryDevtools
                       initialIsOpen={false}
@@ -126,11 +123,9 @@ function MyApp({
                 <I18nextProvider i18n={i18n}>
                   <AnnouncementProvider>
                     <ErrorBoundary FallbackComponent={Error}>
-                      <CrmDataProvider>
-                        <BaseLayout>
-                          <Component {...pageProps} />
-                        </BaseLayout>
-                      </CrmDataProvider>
+                      <BaseLayout>
+                        <Component {...pageProps} />
+                      </BaseLayout>
                     </ErrorBoundary>
                     <AnnouncementWrapper />
                   </AnnouncementProvider>
