+++ frontend/pages/community/crm/companies.tsx
@@ -11,6 +11,7 @@ import { CompanyTable } from "~community/crm/components/organisms/CompanyTable/C
 import TaskModalController from "~community/crm/components/organisms/TaskModalController/TaskModalController";
 import { useCrmStore } from "~community/crm/store/store";
 import { CrmModalTypes } from "~community/crm/types/ModalTypes";
+import { CrmDataProvider } from "~community/crm/v2/providers/CrmDataProvider";
 import useCrmLimitGuard from "~enterprise/crm/hooks/useCrmLimitGuard";
 import { CrmLimitResource } from "~enterprise/crm/types/CrmLimitTypes";
 
@@ -33,31 +34,33 @@ const Companies: NextPage = () => {
   };
 
   return (
-    <ContentLayout
-      breadcrumbs={[
-        { label: translateText(["breadcrumbs", "crm"]) },
-        { label: translateText(["companies", "title"]) }
-      ]}
-      pageHead={translateText(["companies", "pageHead"])}
-      title={translateText(["companies", "title"])}
-      primaryButtonText={translateText(["companies", "addCompanyBtn"])}
-      primaryBtnIconName={IconName.ADD_ICON}
-      onPrimaryButtonClick={onPrimaryButtonClick}
-      isPrimaryBtnLoading={isCheckingCrmLimit}
-      module={Modules.CRM}
-    >
-      <>
-        {selectedCompanyId && (
-          <SidePanelWrapper>
-            <CompanySidePanel />
-          </SidePanelWrapper>
-        )}
+    <CrmDataProvider>
+      <ContentLayout
+        breadcrumbs={[
+          { label: translateText(["breadcrumbs", "crm"]) },
+          { label: translateText(["companies", "title"]) }
+        ]}
+        pageHead={translateText(["companies", "pageHead"])}
+        title={translateText(["companies", "title"])}
+        primaryButtonText={translateText(["companies", "addCompanyBtn"])}
+        primaryBtnIconName={IconName.ADD_ICON}
+        onPrimaryButtonClick={onPrimaryButtonClick}
+        isPrimaryBtnLoading={isCheckingCrmLimit}
+        module={Modules.CRM}
+      >
+        <>
+          {selectedCompanyId && (
+            <SidePanelWrapper>
+              <CompanySidePanel />
+            </SidePanelWrapper>
+          )}
 
-        <CompanyModalController />
-        <TaskModalController />
-        <CompanyTable />
-      </>
-    </ContentLayout>
+          <CompanyModalController />
+          <TaskModalController />
+          <CompanyTable />
+        </>
+      </ContentLayout>
+    </CrmDataProvider>
   );
 };
