+++ frontend/pages/community/crm/tasks.tsx
@@ -11,6 +11,7 @@ import TaskSidePanel from "~community/crm/components/organisms/TaskSidePanel/Tas
 import TasksTable from "~community/crm/components/organisms/TasksTable/TasksTable";
 import { useCrmStore } from "~community/crm/store/store";
 import { CrmModalTypes } from "~community/crm/types/ModalTypes";
+import { CrmDataProvider } from "~community/crm/v2/providers/CrmDataProvider";
 import useCrmLimitGuard from "~enterprise/crm/hooks/useCrmLimitGuard";
 import { CrmLimitResource } from "~enterprise/crm/types/CrmLimitTypes";
 
@@ -53,34 +54,36 @@ const Tasks: NextPage = () => {
   }, []);
 
   return (
-    <ContentLayout
-      breadcrumbs={[
-        { label: translateText(["breadcrumbs", "crm"]) },
-        { label: translateText(["tasks", "title"]) }
-      ]}
-      pageHead={translateText(["tasks", "pageHead"])}
-      title={translateText(["tasks", "title"])}
-      primaryButtonText={translateText(["tasks", "addTaskBtn"])}
-      primaryBtnIconName={IconName.ADD_ICON}
-      containerStyles={{
-        padding: { xs: "1.375rem 2rem 0", lg: "1.375rem 3rem 0" }
-      }}
-      onPrimaryButtonClick={onPrimaryButtonClick}
-      isPrimaryBtnLoading={isCheckingCrmLimit}
-      module={Modules.CRM}
-    >
-      <>
-        {selectedTaskId && (
-          <SidePanelWrapper>
-            <TaskSidePanel />
-          </SidePanelWrapper>
-        )}
-        <div ref={containerRef} className="flex flex-col w-full gap-4">
-          <TaskModalController />
-          <TasksTable />
-        </div>
-      </>
-    </ContentLayout>
+    <CrmDataProvider>
+      <ContentLayout
+        breadcrumbs={[
+          { label: translateText(["breadcrumbs", "crm"]) },
+          { label: translateText(["tasks", "title"]) }
+        ]}
+        pageHead={translateText(["tasks", "pageHead"])}
+        title={translateText(["tasks", "title"])}
+        primaryButtonText={translateText(["tasks", "addTaskBtn"])}
+        primaryBtnIconName={IconName.ADD_ICON}
+        containerStyles={{
+          padding: { xs: "1.375rem 2rem 0", lg: "1.375rem 3rem 0" }
+        }}
+        onPrimaryButtonClick={onPrimaryButtonClick}
+        isPrimaryBtnLoading={isCheckingCrmLimit}
+        module={Modules.CRM}
+      >
+        <>
+          {selectedTaskId && (
+            <SidePanelWrapper>
+              <TaskSidePanel />
+            </SidePanelWrapper>
+          )}
+          <div ref={containerRef} className="flex flex-col w-full gap-4">
+            <TaskModalController />
+            <TasksTable />
+          </div>
+        </>
+      </ContentLayout>
+    </CrmDataProvider>
   );
 };
